requirements.txt
aiogram~=3.17.0
python-dotenv~=1.0.1
flask~=3.1.0
SQLAlchemy~=2.0.36
aiomysql~=0.2.0
cryptography~=44.0.0
python-docx~=1.1.2